# shared/api/bundles/order.openapi.v1.bundle.yaml
openapi: 3.0.3
info:
  title: Order Service
  version: 1.0.0
  description: API для работы с заказами
tags:
  - name: Order
    description: Операции с данными о заказах
paths:
  /api/v1/orders:
    post:
      tags:
        - order
      summary: создание заказа
      operationId: CreateOrderByDetails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_order_request'
      responses:
        '200':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/create_order_response'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/orders/{order_uuid}/pay:
    parameters:
      - $ref: '#/components/parameters/order_uuid'
    post:
      tags:
        - order
      summary: оплата заказа
      operationId: PayOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/pay_order_request'
      responses:
        '200':
          description: payment was successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pay_order_response'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/orders/{order_uuid}:
    parameters:
      - $ref: '#/components/parameters/order_uuid'
    get:
      summary: Get order data by id
      operationId: GetOrderById
      tags:
        - order
      responses:
        '200':
          description: Order information successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/get_order_by_uuid_response'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '404':
          description: Order data not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/orders/{order_uuid}/cancel:
    parameters:
      - $ref: '#/components/parameters/order_uuid'
    post:
      tags:
        - order
      summary: создание заказа
      operationId: deleteOrderByDetails
      responses:
        '204':
          description: No content
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/conflict_error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
components:
  schemas:
    create_order_request:
      type: object
      required:
        - user_uuid
        - part_uuids
      properties:
        user_uuid:
          type: string
          format: uuid
          description: UUID пользователя
          example: 123e4567-e89b-12d3-a456-426614174000
        part_uuids:
          type: array
          description: Список UUID деталей
          minItems: 1
          items:
            type: string
            format: uuid
          example:
            - 550e8400-e29b-41d4-a716-446655440001
            - 550e8400-e29b-41d4-a716-446655440002
    generic_error:
      components:
        schemas:
          Error:
            type: object
            properties:
              code:
                type: integer
                description: HTTP-код ошибки
              message:
                type: string
                description: Описание ошибки
    create_order_response:
      type: object
      required:
        - order_uuid
        - total_price
      properties:
        order_uuid:
          type: string
          description: UUID созданного заказа
          example: 550e8400-e29b-41d4-a716-446655440000
        total_price:
          type: number
          format: double
          description: Итоговая стоимость заказа
          example: 123.45
          minimum: 0
    bad_request_error:
      components:
        schemas:
          Error:
            type: object
            required:
              - code
              - message
            properties:
              code:
                type: integer
                description: HTTP-код ошибки
                example: 400
              message:
                type: string
                description: Описание ошибки
                example: 'Bad Request: Invalid parameter format'
    internal_server_error:
      components:
        schemas:
          Error:
            type: object
            required:
              - code
              - message
            properties:
              code:
                type: integer
                description: HTTP-код ошибки
                example: 500
              message:
                type: string
                description: Описание ошибки
                example: Internal server error occurred
    payment_method:
      type: string
      enum:
        - CARD
        - SBP
        - CREDIT_CARD
        - INVESTOR_MONEY
      description: Способ оплаты
      example: CARD
      x-enum-mapping:
        CARD: Банковская карта
        SBP: Система быстрых платежей
        CREDIT_CARD: Кредитная карта
        INVESTOR_MONEY: Деньги инвестора
    pay_order_request:
      type: object
      required:
        - payment_method
      properties:
        payment_method:
          $ref: '#/components/schemas/payment_method'
    pay_order_response:
      type: object
      required:
        - transaction_uuid
      properties:
        transaction_uuid:
          type: string
          description: uuid транзакции
          example: 123e4567-e89b-12d3-a456-426655440000
    not_found_error:
      components:
        schemas:
          Error:
            type: object
            required:
              - code
              - message
            properties:
              code:
                type: integer
                description: HTTP-код ошибки
                example: 404
              message:
                type: string
                description: Описание ошибки
                example: Order not found
    order_status:
      type: string
      enum:
        - PENDING_PAYMENT
        - PAID
        - CANCELLED
      description: Статусы заказа
      example: PENDING_PAYMENT
      x-enum-descriptions:
        PENDING_PAYMENT: Ожидает оплаты
        PAID: Оплачен
        CANCELLED: Отменён
    get_order_by_uuid_response:
      type: object
      required:
        - order_uuid
        - user_uuid
        - part_uuids
        - total_price
        - status
      properties:
        order_uuid:
          type: string
          format: uuid
          description: Уникальный идентификатор заказа
          example: 550e8400-e29b-41d4-a716-446655440000
        user_uuid:
          type: string
          format: uuid
          description: UUID пользователя
          example: 123e4567-e89b-12d3-a456-426614174000
        part_uuids:
          type: array
          description: Список UUID деталей
          items:
            type: string
            format: uuid
          example:
            - 550e8400-e29b-41d4-a716-446655440001
            - 550e8400-e29b-41d4-a716-446655440002
        total_price:
          type: number
          format: double
          description: Итоговая стоимость
          example: 123.45
          minimum: 0
        transaction_uuid:
          type: string
          format: uuid
          nullable: true
          description: UUID транзакции (если оплачен)
          example: 550e8400-e29b-41d4-a716-446655440003
        payment_method:
          $ref: '#/components/schemas/payment_method'
        status:
          $ref: '#/components/schemas/order_status'
    conflict_error:
      components:
        schemas:
          Error:
            type: object
            required:
              - code
              - message
            properties:
              code:
                type: integer
                description: HTTP-код ошибки
                example: 409
              message:
                type: string
                description: Описание ошибки
                example: Conflict error
  parameters:
    order_uuid:
      name: order_uuid
      in: path
      required: true
      description: Название заказа для которого заправшивается или обновляются данные
      schema:
        type: string
        format: uuid
        example: 123e4567-e89b-12d3-a456-426655440000
x-ogen:
  target: ./shared/pkg/openapi/order/v1
  package: order_v1
  clean: true